gemini_report_writer/requirements.txt
langchain
langgraph
langchain-google-genai
python-dotenv
redis
requests
tavily-python
